Update table size styling for responsive screen
Make table width 90% of the screen width and heighth wise.
Make filters 15% of the container and table 85% of the container.
Move the slippi logo at the top of the filter section. I'm not sure exactly why this is here besides the fact that it looks nice. It could server some function later maybe.